python_scripts/requirements.txt
# Lista de dependencias de Python
# Agrega aquí tus paquetes, por ejemplo:
# requests==2.28.1
Pillow>=9.0.0 # Para procesamiento de imágenes